+++ lib/include/libft.h
@@ -1,15 +1,16 @@
 #ifndef LIBFT_H
 # define LIBFT_H
 
-//	===== includes =====
+//	===== include =====
 # include <stdbool.h>
 # include <stdlib.h>
 # include <unistd.h>
 # include <fcntl.h>
 # include <stdio.h>
 
-# include "types.h"
+# include "type.h"
 # include "string.h"
-# include "colors.h"
+# include "system.h"
+# include "color.h"
 
 #endif
