+++ lib/include/system.h
@@ -1,0 +1,10 @@
+#ifndef SYSTEM_H
+# define SYSTEM_H
+
+# define BUFFER_SIZE 42
+
+# ifndef OPEN_MAX
+#  define OPEN_MAX 256
+# endif
+//@func
+#endif
