--- lib/includes/system.h
@@ -1,7 +1,0 @@
-#ifndef SYSTEM_H
-# define SYSTEM_H
-
-#define BUFFER_SIZE 42
-
-//@func
-#endif
